Update Passenger arrivalTime to a long.

+++ src/Networking/Passenger.java
@@ -7,6 +7,6 @@ package Networking;
  * @param direction The direction a passenger would like to go (UP/DOWN).
  * @param destinationFloor Floor that the passenger would like to arrive at.
  */
-public record Passenger(int arrivalTime, int sourceFloor, Direction direction, int destinationFloor) {
+public record Passenger(long arrivalTime, int sourceFloor, Direction direction, int destinationFloor) {
 
 }
